third_party/WebKit/LayoutTests/http/tests/devtools/indexeddb/upgrade-events-expected.txt

CONSOLE MESSAGE: line 2: InspectorTest.IndexedDB_callback1
CONSOLE MESSAGE: line 2: InspectorTest.IndexedDB_callback2
CONSOLE MESSAGE: line 2: InspectorTest.IndexedDB_callback3
CONSOLE MESSAGE: line 2: InspectorTest.IndexedDB_callback4
CONSOLE MESSAGE: line 2: InspectorTest.IndexedDB_callback5
CONSOLE MESSAGE: line 2: InspectorTest.IndexedDB_callback6
CONSOLE MESSAGE: line 2: InspectorTest.IndexedDB_callback7
CONSOLE MESSAGE: line 2: InspectorTest.IndexedDB_callback8
CONSOLE MESSAGE: line 2: InspectorTest.IndexedDB_callback9
CONSOLE MESSAGE: line 2: InspectorTest.IndexedDB_callback10
Tests that deleted databases do not get recreated.

Preparing database
Verifying that database was created
Loading values
Deleting database
Loading values again (which should fail)
IndexedDBAgent error: Aborted upgrade.
Verifying that database was not recreated
